refactor(experimental): revise transactions schema
