+++ packages/rpc-graphql/src/resolvers/transaction.ts
@@ -1,0 +1,51 @@
+import { SolanaRpcMethods } from '@solana/rpc-core';
+import { Rpc } from '@solana/rpc-transport/dist/types/json-rpc-types';
+
+import { GraphQLCache } from '../cache';
+import { TransactionQueryArgs } from '../schema/transaction';
+
+export async function resolveTransaction(
+    { signature, encoding = 'jsonParsed', ...config }: TransactionQueryArgs,
+    cache: GraphQLCache,
+    rpc: Rpc<SolanaRpcMethods>
+) {
+    const requestConfig = { encoding, ...config };
+
+    const cached = cache.get(signature, requestConfig);
+    if (cached !== null) {
+        return cached;
+    }
+
+    const transaction = await rpc
+        .getTransaction(signature, requestConfig as unknown as Parameters<SolanaRpcMethods['getTransaction']>[1])
+        .send();
+
+    if (transaction === null) {
+        return null;
+    }
+
+    const [transactionData, responseEncoding, responseFormat] = Array.isArray(transaction.transaction)
+        ? encoding === 'jsonParsed'
+            ? [transaction.transaction[0], 'base64', 'unparsed']
+            : [transaction.transaction[0], encoding, 'unparsed']
+        : encoding === 'jsonParsed'
+        ? [transaction.transaction, encoding, 'parsed']
+        : [transaction.transaction, encoding, 'unparsed'];
+    if (transaction.meta) {
+        // Ugly, but tells TypeScript what's happening
+        (transaction.meta as { format?: string } & { [key: string]: unknown })['format'] = responseFormat;
+    }
+    if (transactionData.message) {
+        // Ugly, but tells TypeScript what's happening
+        (transactionData.message as { format?: string } & { [key: string]: unknown })['format'] = responseFormat;
+    }
+    const queryResponse = {
+        ...transaction,
+        encoding: responseEncoding,
+        transaction: transactionData,
+    };
+
+    cache.insert(signature, requestConfig, queryResponse);
+
+    return queryResponse;
+}
